+++ shared/common-adapters/index.js
@@ -74,7 +74,7 @@ export {default as Divider} from './divider'
 export {default as Dropdown} from './dropdown'
 export {default as Emoji} from './emoji'
 export {default as ErrorBoundary} from './error-boundary'
-export {default as FollowButton} from './follow-button'
+export {default as FollowButton} from '../profile/follow-button'
 export {default as FormWithCheckbox} from './form-with-checkbox'
 export {default as Header} from './header'
 export {default as HeaderHoc} from './header-hoc'
@@ -103,7 +103,6 @@ export {default as TabBar} from './tab-bar'
 export {default as Tabs} from './tabs'
 export {default as Text} from './text'
 export {default as TimelineMarker} from './timeline-marker'
-export {default as UserActions} from './user-actions'
 export {default as UserBio} from './user-bio'
 export {default as UserCard} from './user-card'
 export {default as UserProofs} from './user-proofs'
